+++ apps/meteor/client/views/admin/workspace/UsersUploadsCard/UsersUploadsCard.tsx
@@ -1,14 +1,14 @@
 import type { IStats } from '@rocket.chat/core-typings';
-import { ButtonGroup, Button } from '@rocket.chat/fuselage';
+import { Button, Card, CardBody, CardControls } from '@rocket.chat/fuselage';
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
-import { TextSeparator, Card, CardBody, CardCol, CardColSection, CardColTitle, CardFooter, CardIcon } from '@rocket.chat/ui-client';
 import { useRouter, useTranslation } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import React, { memo } from 'react';
 
 import { useHasLicenseModule } from '../../../../../ee/client/hooks/useHasLicenseModule';
-import { UserStatus } from '../../../../components/UserStatus';
 import { useFormatMemorySize } from '../../../../hooks/useFormatMemorySize';
+import WorkspaceCardSection from '../components/WorkspaceCardSection';
+import WorkspaceCardTextSeparator from '../components/WorkspaceCardTextSeparator';
 
 type UsersUploadsCardProps = {
 	statistics: IStats;
@@ -27,81 +27,50 @@ const UsersUploadsCard = ({ statistics }: UsersUploadsCardProps): ReactElement =
 	const canViewEngagement = useHasLicenseModule('engagement-dashboard');
 
 	return (
-		<Card>
-			<CardBody>
-				<CardCol>
-					<CardColSection mbs={0} mbe={16}>
-						<CardColTitle>{t('Users')}</CardColTitle>
-						<TextSeparator
-							label={
-								<>
-									<CardIcon>
-										<UserStatus status='online' />
-									</CardIcon>
-									{t('Online')}
-								</>
-							}
-							value={statistics.onlineUsers}
-						/>
-						<TextSeparator
-							label={
-								<>
-									<CardIcon>
-										<UserStatus status='busy' />
-									</CardIcon>
-									{t('Busy')}
-								</>
-							}
-							value={statistics.busyUsers}
-						/>
-						<TextSeparator
-							label={
-								<>
-									<CardIcon>
-										<UserStatus status='away' />
-									</CardIcon>
-									{t('Away')}
-								</>
-							}
-							value={statistics.awayUsers}
-						/>
-						<TextSeparator
-							label={
-								<>
-									<CardIcon>
-										<UserStatus status='offline' />
-									</CardIcon>
-									{t('Offline')}
-								</>
-							}
-							value={statistics.offlineUsers}
-						/>
-						<TextSeparator label={t('Total')} value={statistics.totalUsers} />
-					</CardColSection>
+		<Card height='full'>
+			<CardBody flexDirection='column'>
+				<WorkspaceCardSection
+					title={t('Users')}
+					body={
+						<>
+							<WorkspaceCardTextSeparator label={t('Online')} status='online' value={statistics.onlineUsers} />
+							<WorkspaceCardTextSeparator label={t('Busy')} status='busy' value={statistics.busyUsers} />
+							<WorkspaceCardTextSeparator label={t('Away')} status='away' value={statistics.awayUsers} />
+							<WorkspaceCardTextSeparator label={t('Offline')} status='offline' value={statistics.offlineUsers} />
+							<WorkspaceCardTextSeparator label={t('Total')} value={statistics.totalUsers} />
+						</>
+					}
+				/>
 
-					<CardColSection mb={16}>
-						<CardColTitle>{t('Types')}</CardColTitle>
-						<TextSeparator label={t('Users_Connected')} value={statistics.totalConnectedUsers} />
-						<TextSeparator label={t('Stats_Active_Users')} value={statistics.activeUsers} />
-						<TextSeparator label={t('Stats_Active_Guests')} value={statistics.activeGuests} />
-						<TextSeparator label={t('Stats_Non_Active_Users')} value={statistics.nonActiveUsers} />
-						<TextSeparator label={t('Stats_App_Users')} value={statistics.appUsers} />
-					</CardColSection>
+				<WorkspaceCardSection
+					title={t('Types')}
+					body={
+						<>
+							<WorkspaceCardTextSeparator label={t('Users_Connected')} value={statistics.totalConnectedUsers} />
+							<WorkspaceCardTextSeparator label={t('Stats_Active_Users')} value={statistics.activeUsers} />
+							<WorkspaceCardTextSeparator label={t('Stats_Active_Guests')} value={statistics.activeGuests} />
+							<WorkspaceCardTextSeparator label={t('Stats_Non_Active_Users')} value={statistics.nonActiveUsers} />
+							<WorkspaceCardTextSeparator label={t('Stats_App_Users')} value={statistics.appUsers} />
+						</>
+					}
+				/>
 
-					<CardColSection mb={16}>
-						<CardColTitle>{t('Uploads')}</CardColTitle>
-						<TextSeparator label={t('Stats_Total_Uploads')} value={statistics.uploadsTotal} />
-						<TextSeparator label={t('Stats_Total_Uploads_Size')} value={formatMemorySize(statistics.uploadsTotalSize)} />
-					</CardColSection>
-				</CardCol>
+				<WorkspaceCardSection
+					title={t('Uploads')}
+					body={
+						<>
+							<WorkspaceCardTextSeparator label={t('Stats_Total_Uploads')} value={statistics.uploadsTotal} />
+							<WorkspaceCardTextSeparator label={t('Stats_Total_Uploads_Size')} value={formatMemorySize(statistics.uploadsTotalSize)} />
+						</>
+					}
+				/>
 			</CardBody>
-			<CardFooter>
-				<ButtonGroup align='end'>
-					<Button disabled={!canViewEngagement} small onClick={handleEngagement}>
-						{t('See_on_Engagement_Dashboard')}
-					</Button>
-				</ButtonGroup>
-			</CardFooter>
+
+			<CardControls>
+				<Button disabled={!canViewEngagement} medium onClick={handleEngagement}>
+					{t('See_on_Engagement_Dashboard')}
+				</Button>
+			</CardControls>
 		</Card>
 	);
 };