+++ apps/meteor/client/views/room/providers/RoomProvider.tsx
@@ -17,7 +17,9 @@ import { useRoomRolesManagement } from '../body/hooks/useRoomRolesManagement';
 import { RoomContext } from '../contexts/RoomContext';
 import ComposerPopupProvider from './ComposerPopupProvider';
 import RoomToolboxProvider from './RoomToolboxProvider';
+import { useRedirectOnSettingsChanged } from './hooks/useRedirectOnSettingsChanged';
 import { useRoomQuery } from './hooks/useRoomQuery';
+import { useUsersNameChanged } from './hooks/useUsersNameChanged';
 
 type RoomProviderProps = {
 	children: ReactNode;
@@ -38,6 +40,10 @@ const RoomProvider = ({ rid, children }: RoomProviderProps): ReactElement => {
 	}, [isSuccess, room, router]);
 
 	const subscriptionQuery = useReactiveQuery(['subscriptions', { rid }], () => ChatSubscription.findOne({ rid }) ?? null);
+
+	useRedirectOnSettingsChanged(subscriptionQuery.data);
+
+	useUsersNameChanged();
 
 	const pseudoRoom = useMemo(() => {
 		if (!room) {
@@ -93,14 +99,7 @@ const RoomProvider = ({ rid, children }: RoomProviderProps): ReactElement => {
 			return;
 		}
 
-		UserAction.addStream(rid);
-		return (): void => {
-			try {
-				UserAction.cancel(rid);
-			} catch (error) {
-				// Do nothing
-			}
-		};
+		return UserAction.addStream(rid);
 	}, [rid, subscribed]);
 
 	if (!pseudoRoom) {