+++ apps/meteor/ee/client/omnichannel/reports/sections/ChannelsSection.tsx
@@ -9,7 +9,7 @@ export const ChannelsSection = () => {
 	const colors = useMemo(() => Object.values(COLORS), []);
 
 	return (
-		<ReportCard {...config} minHeight={200}>
+		<ReportCard {...config}>
 			<PieChart data={data} width={200} height={200} colors={colors} />
 		</ReportCard>
 	);