+++ packages/uikit-playground/src/Context/action/duplicateProjectAction.ts
@@ -1,0 +1,14 @@
+import { ActionTypes } from "../reducer";
+
+export type DuplicateProjectAction = {
+  type: ActionTypes.DuplicateProject,
+  payload: { id: string, name?: string },
+};
+
+export const duplicateProjectAction = (payload: {
+  id: string,
+  name?: string,
+}): DuplicateProjectAction => ({
+  type: ActionTypes.DuplicateProject,
+  payload,
+});